+++ test/MicroElements.Metadata.Tests/SchemaTests/SimpleSchemaTests.cs
@@ -28,29 +28,29 @@ namespace MicroElements.Metadata.Tests.SchemaTests
 
             Action setDefaultValue = () => schema.SetDefaultValueUntyped(null);
             setDefaultValue.Should().NotThrow();
-            schema.GetDefaultValue().Should().NotBeNull();
-            schema.GetDefaultValue().GetDefaultValue().Should().BeNull();
+            schema.GetDefaultValueMetadata().Should().NotBeNull();
+            schema.GetDefaultValueMetadata().Value.Should().BeNull();
 
             schema = new SimpleTypeSchema("Currency", typeof(string));
             setDefaultValue = () => schema.SetDefaultValueUntyped(1);
-            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("Value 1 can not be set as default value for type System.String");
-            schema.GetDefaultValue().Should().BeNull();
+            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("Value '1' can not be set as default value for type 'System.String'");
+            schema.GetDefaultValueMetadata().Should().BeNull();
 
             schema = new SimpleTypeSchema("Money", typeof(double));
             setDefaultValue = () => schema.SetDefaultValueUntyped(0.0);
             setDefaultValue.Should().NotThrow();
-            schema.GetDefaultValue().Should().NotBeNull();
-            schema.GetDefaultValue().GetDefaultValue().Should().Be(0.0);
+            schema.GetDefaultValueMetadata().Should().NotBeNull();
+            schema.GetDefaultValueMetadata().Value.Should().Be(0.0);
 
             schema = new SimpleTypeSchema("Money", typeof(double));
             setDefaultValue = () => schema.SetDefaultValueUntyped(100);
-            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("Value 100 can not be set as default value for type System.Double");
-            schema.GetDefaultValue().Should().BeNull();
+            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("Value '100' can not be set as default value for type 'System.Double'");
+            schema.GetDefaultValueMetadata().Should().BeNull();
 
             schema = new SimpleTypeSchema("Money", typeof(double));
             setDefaultValue = () => schema.SetDefaultValueUntyped(null);
-            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("null value can not be set as default value for type System.Double");
-            schema.GetDefaultValue().Should().BeNull();
+            setDefaultValue.Should().Throw<Exception>().And.Message.Should().Be("Value 'null' can not be set as default value for type 'System.Double'");
+            schema.GetDefaultValueMetadata().Should().BeNull();
         }
 
         [Fact]
@@ -60,32 +60,38 @@ namespace MicroElements.Metadata.Tests.SchemaTests
                 var schema = new SimpleTypeSchema<string>("Currency");
                 Action setDefaultValue = () => schema.SetDefaultValue(null);
                 setDefaultValue.Should().NotThrow();
-                schema.GetDefaultValue().Should().NotBeNull();
-                schema.GetDefaultValue().GetDefaultValue().Should().Be(null);
+                schema.GetDefaultValueMetadata().Should().NotBeNull();
+                schema.GetDefaultValueMetadata().Value.Should().Be(null);
             }
 
             {
                 var schema = new SimpleTypeSchema<string>("Currency");
                 Action setDefaultValue = () => schema.SetDefaultValue("empty");
                 setDefaultValue.Should().NotThrow();
-                schema.GetDefaultValue().Should().NotBeNull();
-                schema.GetDefaultValue().GetDefaultValue().Should().Be("empty");
+                schema.GetDefaultValueMetadata().Should().NotBeNull();
+                schema.GetDefaultValueMetadata().Value.Should().Be("empty");
             }
 
             {
                 var schema = new SimpleTypeSchema<double>("Money");
                 Action setDefaultValue = () => schema.SetDefaultValue(0.0);
                 setDefaultValue.Should().NotThrow();
-                schema.GetDefaultValue().Should().NotBeNull();
-                schema.GetDefaultValue().GetDefaultValue().Should().Be(0.0);
+                schema.GetDefaultValueMetadata().Should().NotBeNull();
+                schema.GetDefaultValueMetadata().Value.Should().Be(0.0);
             }
 
             {
                 var schema = new SimpleTypeSchema<double>("Money");
                 Action setDefaultValue = () => schema.SetDefaultValue(100);
                 setDefaultValue.Should().NotThrow();
-                schema.GetDefaultValue().Should().NotBeNull();
-                schema.GetDefaultValue().GetDefaultValue().Should().Be(100.0);
+                schema.GetDefaultValueMetadata().Should().NotBeNull();
+                schema.GetDefaultValueMetadata().Value.Should().Be(100.0);
+            }
+
+            {
+                var schema = new Property<double>("Money").SetDefaultValue(100);
+                schema.GetDefaultValueMetadata().Should().NotBeNull();
+                schema.GetDefaultValue().Should().Be(100.0);
             }
         }